backend/requirements/common.txt
django==4.1.2
Pillow==9.2.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
pydenticon==0.3.1
